# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseError = {
  code: string;
  message: string;
};

export interface ErrorResponse {
  success: boolean;
  error: ErrorResponseError;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Service {
  id: string;
  slug: string;
  name: string;
  description: string;
  shortDescription: string;
  icon: string;
  category: string;
  startingPrice: number;
  features: string[];
}

export interface ServicesResponse {
  success: boolean;
  data: Service[];
}

export interface ServiceDetailResponse {
  success: boolean;
  data: Service;
}

export interface PricingItem {
  item: string;
  type?: string;
  price: number;
  unit?: string;
}

export interface PricingCategory {
  category: string;
  items: PricingItem[];
}

export interface PricingResponse {
  success: boolean;
  data: PricingCategory[];
}

export type SchedulePickupRequestTimeSlot =
  (typeof SchedulePickupRequestTimeSlot)[keyof typeof SchedulePickupRequestTimeSlot];

export const SchedulePickupRequestTimeSlot = {
  morning: "morning",
  afternoon: "afternoon",
  evening: "evening",
} as const;

export type SchedulePickupRequestBranch =
  (typeof SchedulePickupRequestBranch)[keyof typeof SchedulePickupRequestBranch];

export const SchedulePickupRequestBranch = {
  pollachi: "pollachi",
  kinathukadavu: "kinathukadavu",
} as const;

export interface SchedulePickupRequest {
  name: string;
  phone: string;
  address: string;
  addressLat?: number;
  addressLng?: number;
  services: string[];
  specialInstructions?: string;
  preferredDate: string;
  timeSlot: SchedulePickupRequestTimeSlot;
  branch: SchedulePickupRequestBranch;
}

export type SchedulePickupResponseData = {
  bookingReference: string;
  message: string;
};

export interface SchedulePickupResponse {
  success: boolean;
  data: SchedulePickupResponseData;
}

export interface ContactRequest {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface SendOtpRequest {
  phone: string;
}

export type SendOtpResponseData = {
  message: string;
  expiresIn: number;
};

export interface SendOtpResponse {
  success: boolean;
  data: SendOtpResponseData;
}

export interface VerifyOtpRequest {
  phone: string;
  otp: string;
}

export interface UserProfile {
  id: string;
  phone: string;
  name?: string | null;
  email?: string | null;
  isActive?: boolean;
  createdAt?: string | null;
}

export type AuthResponseData = {
  accessToken: string;
  isNewUser: boolean;
  user: UserProfile;
};

export interface AuthResponse {
  success: boolean;
  data: AuthResponseData;
}

export type RefreshResponseData = {
  accessToken: string;
};

export interface RefreshResponse {
  success: boolean;
  data: RefreshResponseData;
}

export interface ProfileResponse {
  success: boolean;
  data: UserProfile;
}

export interface UpdateProfileRequest {
  name?: string;
  email?: string;
}

export interface PortalOrderItem {
  serviceName: string;
  quantity: number;
  price: number;
}

export interface TrackingStep {
  key: string;
  label: string;
  completed: boolean;
  current: boolean;
}

export type PortalOrderStatus =
  (typeof PortalOrderStatus)[keyof typeof PortalOrderStatus];

export const PortalOrderStatus = {
  pending: "pending",
  processing: "processing",
  ready_for_pickup: "ready_for_pickup",
  out_for_delivery: "out_for_delivery",
  completed: "completed",
  delivered: "delivered",
  in_store: "in_store",
  ready: "ready",
  cancelled: "cancelled",
} as const;

export type PortalOrderFulfillmentType =
  (typeof PortalOrderFulfillmentType)[keyof typeof PortalOrderFulfillmentType];

export const PortalOrderFulfillmentType = {
  pickup: "pickup",
  delivery: "delivery",
} as const;

export interface PortalOrder {
  id: string;
  orderNumber: string;
  reference: string;
  status: PortalOrderStatus;
  paymentStatus: string;
  totalAmount?: number | null;
  services: string[];
  items: PortalOrderItem[];
  branch: string;
  fulfillmentType: PortalOrderFulfillmentType;
  scheduledDate?: string | null;
  createdAt: string;
  updatedAt?: string | null;
  pickupDate?: string | null;
  invoiceUrl?: string | null;
  lastWhatsappStatus?: string | null;
  lastWhatsappSentAt?: string | null;
  customerName?: string | null;
  customerPhone?: string | null;
  customerEmail?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  customerRating?: number | null;
  feedbackComment?: string | null;
  feedbackSubmittedAt?: string | null;
}

export interface OrdersResponse {
  success: boolean;
  data: PortalOrder[];
}

export interface OrderDetailResponse {
  success: boolean;
  data: PortalOrder;
}

export type PublicTrackOrder = PortalOrder & {
  steps: TrackingStep[];
};

export interface PublicTrackOrderResponse {
  success: boolean;
  data: PublicTrackOrder;
}

export interface ExistingFeedback {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number | null;
  comment: string | null;
  submittedAt: string | null;
}

export type FeedbackContextFulfillmentType =
  (typeof FeedbackContextFulfillmentType)[keyof typeof FeedbackContextFulfillmentType];

export const FeedbackContextFulfillmentType = {
  pickup: "pickup",
  delivery: "delivery",
} as const;

export interface FeedbackContext {
  id: string;
  orderNumber: string;
  customerName?: string | null;
  status: string;
  fulfillmentType: FeedbackContextFulfillmentType;
  totalAmount?: number | null;
  items: PortalOrderItem[];
  existingFeedback: ExistingFeedback;
}

export interface FeedbackContextResponse {
  success: boolean;
  data: FeedbackContext;
}

export type SubmitFeedbackRequestMetadata = { [key: string]: unknown };

export interface SubmitFeedbackRequest {
  orderId?: string;
  orderNumber?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
  metadata?: SubmitFeedbackRequestMetadata;
}

export interface FeedbackSubmission {
  orderId: string;
  orderNumber: string;
  rating: number;
  comment?: string | null;
  feedbackSubmittedAt: string;
  googleReviewEligible: boolean;
}

export interface SubmitFeedbackResponse {
  success: boolean;
  data: FeedbackSubmission;
}

export interface PublicReview {
  id: string;
  rating: number;
  comment?: string | null;
  isTopRating: boolean;
  isBestRating: boolean;
  curationScore?: number | null;
  curationReason?: string | null;
  aiProvider?: string | null;
  aiModel?: string | null;
  createdAt: string;
  customerName: string;
  location: string;
  orderNumber?: string | null;
}

export interface TopReviewsResponse {
  success: boolean;
  data: PublicReview[];
}

export interface BestReviewsMeta {
  page: number;
  pageSize: number;
  hasMore: boolean;
}

export interface BestReviewsResponse {
  success: boolean;
  data: PublicReview[];
  meta: BestReviewsMeta;
}

export type GetFeedbackContextParams = {
  orderId?: string;
  orderNumber?: string;
};

export type ListBestReviewsParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 50
   */
  pageSize?: number;
};