# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fab Clean API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List portal user orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get portal order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /track/{orderNumber}:
    get:
      operationId: getPublicTrackOrder
      tags: [tracking]
      summary: Track an ERP order by order number
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ERP-backed order tracking data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTrackOrderResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feedback/context:
    get:
      operationId: getFeedbackContext
      tags: [feedback]
      summary: Load feedback context for an ERP order
      parameters:
        - name: orderId
          in: query
          required: false
          schema:
            type: string
        - name: orderNumber
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Feedback context loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackContextResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit or update ERP-backed order feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackRequest"
      responses:
        "200":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitFeedbackResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reviews/best:
    get:
      operationId: listBestReviews
      tags: [reviews]
      summary: List paginated best reviews
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
      responses:
        "200":
          description: Paginated best reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BestReviewsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
    PortalOrderItem:
      type: object
      properties:
        serviceName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required: [serviceName, quantity, price]
    TrackingStep:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        completed:
          type: boolean
        current:
          type: boolean
      required: [key, label, completed, current]
    PortalOrder:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        reference:
          type: string
        status:
          type: string
          enum:
            - pending
            - processing
            - ready_for_pickup
            - out_for_delivery
            - completed
            - delivered
            - in_store
            - ready
            - cancelled
        paymentStatus:
          type: string
        totalAmount:
          type: number
          nullable: true
        services:
          type: array
          items:
            type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PortalOrderItem"
        branch:
          type: string
        fulfillmentType:
          type: string
          enum: [pickup, delivery]
        scheduledDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        pickupDate:
          type: string
          format: date-time
          nullable: true
        invoiceUrl:
          type: string
          nullable: true
        lastWhatsappStatus:
          type: string
          nullable: true
        lastWhatsappSentAt:
          type: string
          format: date-time
          nullable: true
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        customerEmail:
          type: string
          nullable: true
        customerRating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        feedbackComment:
          type: string
          nullable: true
        feedbackSubmittedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - orderNumber
        - reference
        - status
        - paymentStatus
        - services
        - items
        - branch
        - fulfillmentType
        - createdAt
    OrdersResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/PortalOrder"
      required: [success, data]
    OrderDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/PortalOrder"
      required: [success, data]
    PublicTrackOrder:
      allOf:
        - $ref: "#/components/schemas/PortalOrder"
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: "#/components/schemas/TrackingStep"
          required: [steps]
    PublicTrackOrderResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/PublicTrackOrder"
      required: [success, data]
    ExistingFeedback:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        comment:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
      required: [rating, comment, submittedAt]
    FeedbackContext:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        customerName:
          type: string
          nullable: true
        status:
          type: string
        fulfillmentType:
          type: string
          enum: [pickup, delivery]
        totalAmount:
          type: number
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PortalOrderItem"
        existingFeedback:
          $ref: "#/components/schemas/ExistingFeedback"
      required:
        - id
        - orderNumber
        - status
        - fulfillmentType
        - items
        - existingFeedback
    FeedbackContextResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/FeedbackContext"
      required: [success, data]
    SubmitFeedbackRequest:
      type: object
      properties:
        orderId:
          type: string
        orderNumber:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
      required: [rating]
    FeedbackSubmission:
      type: object
      properties:
        orderId:
          type: string
        orderNumber:
          type: string
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        feedbackSubmittedAt:
          type: string
          format: date-time
        googleReviewEligible:
          type: boolean
      required:
        - orderId
        - orderNumber
        - rating
        - feedbackSubmittedAt
        - googleReviewEligible
    SubmitFeedbackResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/FeedbackSubmission"
      required: [success, data]
    PublicReview:
      type: object
      properties:
        id:
          type: string
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        isTopRating:
          type: boolean
        isBestRating:
          type: boolean
        curationScore:
          type: number
          nullable: true
        curationReason:
          type: string
          nullable: true
        aiProvider:
          type: string
          nullable: true
        aiModel:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        customerName:
          type: string
        location:
          type: string
        orderNumber:
          type: string
          nullable: true
      required:
        - id
        - rating
        - isTopRating
        - isBestRating
        - createdAt
        - customerName
        - location
...
    BestReviewsMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean
      required: [page, pageSize, hasMore]
    BestReviewsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/PublicReview"
        meta:
          $ref: "#/components/schemas/BestReviewsMeta"
      required: [success, data, meta]