# lib/api-zod/src/generated/types/bestReviewsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BestReviewsMeta } from "./bestReviewsMeta";
import type { PublicReview } from "./publicReview";

export interface BestReviewsResponse {
  success: boolean;
  data: PublicReview[];
  meta: BestReviewsMeta;
}