# lib/api-zod/src/generated/types/existingFeedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExistingFeedback {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number | null;
  comment: string | null;
  submittedAt: Date | null;
}