# lib/api-zod/src/generated/types/feedbackContext.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExistingFeedback } from "./existingFeedback";
import type { FeedbackContextFulfillmentType } from "./feedbackContextFulfillmentType";
import type { PortalOrderItem } from "./portalOrderItem";

export interface FeedbackContext {
  id: string;
  orderNumber: string;
  customerName?: string | null;
  status: string;
  fulfillmentType: FeedbackContextFulfillmentType;
  totalAmount?: number | null;
  items: PortalOrderItem[];
  existingFeedback: ExistingFeedback;
}