# lib/api-zod/src/generated/types/feedbackSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackSubmission {
  orderId: string;
  orderNumber: string;
  rating: number;
  comment?: string | null;
  feedbackSubmittedAt: Date;
  googleReviewEligible: boolean;
}