# lib/api-zod/src/generated/types/portalOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PortalOrderFulfillmentType } from "./portalOrderFulfillmentType";
import type { PortalOrderItem } from "./portalOrderItem";
import type { PortalOrderStatus } from "./portalOrderStatus";

export interface PortalOrder {
  id: string;
  orderNumber: string;
  reference: string;
  status: PortalOrderStatus;
  paymentStatus: string;
  totalAmount?: number | null;
  services: string[];
  items: PortalOrderItem[];
  branch: string;
  fulfillmentType: PortalOrderFulfillmentType;
  scheduledDate?: Date | null;
  createdAt: Date;
  updatedAt?: Date | null;
  pickupDate?: Date | null;
  invoiceUrl?: string | null;
  lastWhatsappStatus?: string | null;
  lastWhatsappSentAt?: Date | null;
  customerName?: string | null;
  customerPhone?: string | null;
  customerEmail?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  customerRating?: number | null;
  feedbackComment?: string | null;
  feedbackSubmittedAt?: Date | null;
}