# lib/api-zod/src/generated/types/portalOrderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortalOrderItem {
  serviceName: string;
  quantity: number;
  price: number;
}