# lib/api-zod/src/generated/types/publicReview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicReview {
  id: string;
  rating: number;
  comment?: string | null;
  isTopRating: boolean;
  isBestRating: boolean;
  curationScore?: number | null;
  curationReason?: string | null;
  aiProvider?: string | null;
  aiModel?: string | null;
  createdAt: Date;
  customerName: string;
  location: string;
  orderNumber?: string | null;
}