# lib/api-zod/src/generated/types/publicTrackOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PortalOrder } from "./portalOrder";
import type { TrackingStep } from "./trackingStep";

export type PublicTrackOrder = PortalOrder & {
  steps: TrackingStep[];
};