# lib/api-zod/src/generated/types/publicTrackOrderResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicTrackOrder } from "./publicTrackOrder";

export interface PublicTrackOrderResponse {
  success: boolean;
  data: PublicTrackOrder;
}