# lib/api-zod/src/generated/types/trackingStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fab Clean API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrackingStep {
  key: string;
  label: string;
  completed: boolean;
  current: boolean;
}